Migrate entire project into a bundler via Vite
