--- README.md
@@ -1,2 +1,0 @@
-# weather-app
-A weather app made using plain JavaScript and the Visual Crossing API.
