+++ index.html
@@ -5,13 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Weather App</title>
-    <!-- For the Pretendard font -->
-    <link rel="stylesheet" href="font/pretendard.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/reset.css">
-    <link rel="stylesheet" href="css/modal.css">
-
-    <script src="code/index.js" type="module"></script>
+    <script src="src/main.js" type="module"></script>
 </head>
 
 <body>
