+++ src/code/weatherManager.js
