+++ src/main.js
@@ -1,0 +1,16 @@
+import WeatherManager from "./code/weatherManager.js";
+import DOMManager from "./code/domManager.js";
+import DateManager from "./code/dateManager.js";
+
+// Since I'm using Vite, let's import them that way
+import "./css/reset.css";
+import "./font/pretendard.css";
+import "./css/modal.css";
+import "./css/style.css";
+
+await WeatherManager.initialize("Maple Ridge, BC");
+await DOMManager.initialize();
+
+WeatherManager.logWeatherData();
+const dateforDay2 = WeatherManager.getWeekForecast()[2][0];
+console.log(`The date is: ${dateforDay2} and the day is: ${DateManager.getDayOfWeek(dateforDay2)}`);
